services/weatherService/requirements.txt
pyzmq==27.1.0
requests==2.33.1
python-dotenv==1.2.2